refactor/fontColor
