+++ src/components/nav/nav.module.css
@@ -18,7 +18,7 @@
     align-items: center;
     justify-content: center;
     text-decoration: none;
-    color: black;
+    color: #767676;
     font-size: 10px;
     font-weight: 400;
 }
